pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icb-worker"
version = "0.1.0"
description = "Coffee scraping worker and orchestrator"
authors = [{name = "ICB Team", email = "team@indiancoffeebeans.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
    "bullmq>=0.1.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
